impl: renamed the package to log from logger
