+++ log/loggcp/logger_gcp.go
